Dev container for ROS2 Jazzy on Ubuntu Noble
Was getting some strange errors on the osrf ROS docker container,
so I switched to the official docker hub one for now.

Note: The use of CMAKE_PREFIX_PATH seems to have changed. As long as
AMENT_PREFIX_PATH is set, it seems that colcon must internally set the
CMAKE_PREFIX_PATH for cmake.